remove database drop/recreate
Dropping tables not needed as it can cause issues when changing it in use, everything
can be added with a separate application.

Indexes dropped as they don't have a IF NOT EXISTS clause and are hardly
worth it for the small ammount of data.

Remove auto inserting of domains,aliases and mail addresses this is will
also be added with an external application.